Changed structure of "Bank.cs" so that the user have to login before they see the user menu.
Added a condition to "login.cs" that checks if the user is admin or not, if they are admin they will be sent to the "CreateAccount"´method in "account" class, if they are a user they will see the user menu.

Removed method "adminLogin()" in login.cs becuase i decided to check if the user is admin with the condition mentioned above.